Worked on end game cases, waters rise cards put back into deck
The code was looping infinitely since the code wasnt setup to use a sequence of lift cards to leave the island. Worked on this.
The waters rise cards were not being put back into the discard pile. They are now

There is some kind of issue with this code. The win rate has dropped to nearly 0%. It's most likely an issue with the new conditions intended to help win the game when close to victory